fix(redis): use new tracer after setTracerProvider
