+++ plugins/node/opentelemetry-instrumentation-redis/src/instrumentation.ts
@@ -18,16 +18,28 @@ import {
   isWrapped,
   InstrumentationBase,
   InstrumentationNodeModuleDefinition,
+  safeExecuteInTheMiddle,
 } from '@opentelemetry/instrumentation';
 import {
+    endSpan,
   getTracedCreateClient,
   getTracedCreateStreamTrace,
-  getTracedInternalSendCommand,
 } from './utils';
-import { RedisInstrumentationConfig } from './types';
+import { RedisCommand, RedisInstrumentationConfig } from './types';
 /** @knipignore */
 import { PACKAGE_NAME, PACKAGE_VERSION } from './version';
-
+import { RedisPluginClientTypes } from './internal-types';
+import { SpanKind, context, trace } from '@opentelemetry/api';
+import {
+    DBSYSTEMVALUES_REDIS,
+    SEMATTRS_DB_CONNECTION_STRING,
+    SEMATTRS_DB_STATEMENT,
+    SEMATTRS_DB_SYSTEM,
+    SEMATTRS_NET_PEER_NAME,
+    SEMATTRS_NET_PEER_PORT,
+  } from '@opentelemetry/semantic-conventions';
+import { defaultDbStatementSerializer } from '@opentelemetry/redis-common';
+  
 const DEFAULT_CONFIG: RedisInstrumentationConfig = {
   requireParentSpan: false,
 };
@@ -96,28 +108,110 @@ export class RedisInstrumentation extends InstrumentationBase<RedisInstrumentati
       ),
     ];
   }
+
   /**
    * Patch internal_send_command(...) to trace requests
    */
   private _getPatchInternalSendCommand() {
-    const tracer = this.tracer;
-    const config = this.getConfig();
+    const instrumentation = this;
     return function internal_send_command(original: Function) {
-      return getTracedInternalSendCommand(tracer, original, config);
+        return function internal_send_command_trace(
+            this: RedisPluginClientTypes,
+            cmd?: RedisCommand
+          ) {
+            // Versions of redis (2.4+) use a single options object
+            // instead of named arguments
+            if (arguments.length !== 1 || typeof cmd !== 'object') {
+              // We don't know how to trace this call, so don't start/stop a span
+              return original.apply(this, arguments);
+            }
+
+            const config = instrumentation.getConfig();
+        
+            const hasNoParentSpan = trace.getSpan(context.active()) === undefined;
+            if (config.requireParentSpan === true && hasNoParentSpan) {
+              return original.apply(this, arguments);
+            }
+        
+            const dbStatementSerializer =
+              config?.dbStatementSerializer || defaultDbStatementSerializer;
+            const span = instrumentation.tracer.startSpan(
+              `${RedisInstrumentation.COMPONENT}-${cmd.command}`,
+              {
+                kind: SpanKind.CLIENT,
+                attributes: {
+                  [SEMATTRS_DB_SYSTEM]: DBSYSTEMVALUES_REDIS,
+                  [SEMATTRS_DB_STATEMENT]: dbStatementSerializer(cmd.command, cmd.args),
+                },
+              }
+            );
+        
+            // Set attributes for not explicitly typed RedisPluginClientTypes
+            if (this.connection_options) {
+              span.setAttributes({
+                [SEMATTRS_NET_PEER_NAME]: this.connection_options.host,
+                [SEMATTRS_NET_PEER_PORT]: this.connection_options.port,
+              });
+            }
+            if (this.address) {
+              span.setAttribute(
+                SEMATTRS_DB_CONNECTION_STRING,
+                `redis://${this.address}`
+              );
+            }
+        
+            const originalCallback = arguments[0].callback;
+            if (originalCallback) {
+              const originalContext = context.active();
+              (arguments[0] as RedisCommand).callback = function callback<T>(
+                this: unknown,
+                err: Error | null,
+                reply: T
+              ) {
+                if (config?.responseHook) {
+                  const responseHook = config.responseHook;
+                  safeExecuteInTheMiddle(
+                    () => {
+                      responseHook(span, cmd.command, cmd.args, reply);
+                    },
+                    err => {
+                      if (err) {
+                        instrumentation._diag.error('Error executing responseHook', err);
+                      }
+                    },
+                    true
+                  );
+                }
+        
+                endSpan(span, err);
+                return context.with(
+                  originalContext,
+                  originalCallback,
+                  this,
+                  ...arguments
+                );
+              };
+            }
+            try {
+              // Span will be ended in callback
+              return original.apply(this, arguments);
+            } catch (rethrow: any) {
+              endSpan(span, rethrow);
+              throw rethrow; // rethrow after ending span
+            }
+          };
     };
   }
 
   private _getPatchCreateClient() {
-    const tracer = this.tracer;
     return function createClient(original: Function) {
-      return getTracedCreateClient(tracer, original);
+      return getTracedCreateClient(original);
     };
   }
 
   private _getPatchCreateStream() {
-    const tracer = this.tracer;
     return function createReadStream(original: Function) {
-      return getTracedCreateStreamTrace(tracer, original);
+      return getTracedCreateStreamTrace(original);
     };
   }
 }
